% Auto-generated: do not edit by hand
\name{basicUI}

\alias{basicUI}

\title{BasicUI component}

\description{

}

\usage{
basicUI(id=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

