% Auto-generated: do not edit by hand
\name{darkUI}

\alias{darkUI}

\title{DarkUI component}

\description{
Advanced UI in Dark mode
}

\usage{
darkUI(id=NULL, log=NULL, mapboxAccessToken=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{log}{Named list. A string representing the logs}

\item{mapboxAccessToken}{Character. Mapbox API token}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

