% Auto-generated: do not edit by hand
\name{lightUI}

\alias{lightUI}

\title{LightUI component}

\description{

}

\usage{
lightUI(id=NULL, log=NULL, mapboxAccessToken=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{log}{Named list. A string representing the logs}

\item{mapboxAccessToken}{Named list. Mapbox API token}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

