% Auto-generated: do not edit by hand
\name{veryBasicUI}

\alias{veryBasicUI}

\title{VeryBasicUI component}

\description{

}

\usage{
veryBasicUI(id=NULL, log=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{log}{Named list. Temp}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

