% Auto-generated: do not edit by hand
\name{basicUI}

\alias{basicUI}

\title{BasicUI component}

\description{

}

\usage{
basicUI(id=NULL, log=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{log}{Named list. A string representing the logs}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

